fix #25 (#31)
* Change AutoFiled and SmallAutoField to clickhouse Int64, so that id worker can generate value for them.

* `DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'` is no longer a required configuration item.

+++ clickhouse_backend/backend/base.py
@@ -23,8 +23,8 @@ class DatabaseWrapper(BaseDatabaseWrapper):
     # If a column type is set to None, it won't be included in the output.
     data_types = {
         # Django fields.
-        "SmallAutoField": "Int16",
-        "AutoField": "Int32",
+        "SmallAutoField": "Int64",
+        "AutoField": "Int64",
         "BigAutoField": "Int64",
         "IPAddressField": "IPv4",
         "GenericIPAddressField": "IPv6",
